Big boi tutorial update
+ A scripted, with dialogue, tutorial is now present under the "Tutorial" level
+ Tutorial presents two versions depending on InputType setting
+ A duplicate of 'SampleScene' without Tutorial Balls can be found under "Level 1" now

Currently working on:
- The English language

Current bug report:
Dialogues for "FirstHit" and "firstDeath" require the user to click on the screen first to register the eventSystem before being able to click the button. This behaviour is not observed for "firstLoad" or "firstWin" events. "firstLoad" may contain some events that is not visible as the ball is instantiated and "firstWin" contains events pertaining to the Results Screen. Perhaps look at "firstWin" and see if a solution could be created based on its code path.